+++ apps/server/src/api/transactions/stats/index.ts
@@ -1,5 +1,6 @@
 import { z } from "zod";
 import { sql } from "~/db/db";
+import { apiCacheConfig } from "~/lib/api";
 import { getValidatedQueryZod } from "~/lib/nitro";
 import { protocols } from "~/lib/protocols";
 
@@ -7,12 +8,12 @@ const transactionStatsRouteSchema = z.object({
   protocol: z.enum(protocols).optional(),
 });
 
-export type TransactionStatsRouteResponse = {
+type TransactionStatsRouteResponse = {
   count: number;
   unique_senders: number;
 };
 
-export default defineEventHandler(async (event) => {
+export default defineCachedEventHandler(async (event) => {
   const query = await getValidatedQueryZod(event, transactionStatsRouteSchema);
 
   let protocolCondition = "";
@@ -39,4 +40,4 @@ WHERE
   };
 
   return stats;
-});
+}, apiCacheConfig);