+++ apps/web/src/app/protocols/[protocol]/page.tsx
@@ -1,37 +1,16 @@
-import { ProtocolInfo } from "@/components/Protocol/ProtocolInfo";
 import { ProtocolStats } from "@/components/Protocol/ProtocolStats";
 import { ProtocolTransactions } from "@/components/Protocol/ProtocolTransactions";
 import { StackingDAORef } from "@/components/Protocol/StackingDAO/StackingDAORef";
 import { UniqueUsersBarChart } from "@/components/Stats/UniqueUsersBarChart";
 import { DepositWithdrawBarChart } from "@/components/Stats/stackingdao/DepositsWithdrawBarChart";
 import { isProtocol, protocolsInfo } from "@/lib/protocols";
-import { Container } from "@radix-ui/themes";
-import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { Suspense } from "react";
 
 export const dynamic = "force-dynamic";
-// TODO
-// Cache the page for 60 seconds
-// export const revalidate = 60;
 
 interface PageProps {
   params: { protocol: string };
-}
-
-export function generateMetadata({ params }: PageProps): Metadata {
-  const protocol = params.protocol;
-  if (!isProtocol(protocol)) {
-    notFound();
-  }
-
-  return {
-    title: `stackspulse - ${protocolsInfo[protocol].name}`,
-    description: `Get the latest on-chain stats for ${protocolsInfo[protocol].name}. Explore real-time feed, unique users, transactions, and more..`,
-    alternates: {
-      canonical: `/protocols/${protocol}`,
-    },
-  };
 }
 
 export default async function ProtocolPage({ params }: PageProps) {
@@ -40,10 +19,10 @@ export default async function ProtocolPage({ params }: PageProps) {
     notFound();
   }
 
-  return (
-    <Container size="2" className="px-4 pt-10">
-      <ProtocolInfo protocol={protocol} />
+  const protocolInfo = protocolsInfo[protocol];
 
+  return (
+    <>
       {protocol === "stackingdao" ? <StackingDAORef /> : null}
 
       <Suspense>
@@ -63,6 +42,6 @@ export default async function ProtocolPage({ params }: PageProps) {
       <Suspense>
         <ProtocolTransactions protocol={protocol} />
       </Suspense>
-    </Container>
+    </>
   );
 }