+++ apps/web/src/components/Transaction/TransactionRow.tsx
@@ -1,20 +1,20 @@
-import type { SelectTransactionAction } from "@/db/transactions";
+import type { TransactionsRouteResponse } from "@/lib/api";
 import { Button, IconButton, Text, Tooltip } from "@radix-ui/themes";
 import { IconExternalLink } from "@tabler/icons-react";
 import Image from "next/image";
 import Link from "next/link";
 import { TimeAgo } from "../Shared/TimeAgo";
-import { TransactionActionAddLiquidity } from "./Action/AddLiquidity";
-import { TransactionActionRemoveLiquidity } from "./Action/RemoveLiquidity";
+import { TransactionActionAlex } from "./Action/Alex";
+import { DefaultTransactionAction } from "./Action/DefaultTransactionAction";
 import { TransactionActionStackingDAO } from "./Action/StackingDAO";
-import { TransactionActionSwap } from "./Action/Swap";
+import { TransactionActionVelar } from "./Action/Velar";
 
 interface TransactionRowProps {
-  transaction: SelectTransactionAction;
+  transaction: TransactionsRouteResponse[number];
 }
 
 export const TransactionRow = ({ transaction }: TransactionRowProps) => {
-  const timestamp = new Date(transaction.timestamp);
+  const timestamp = new Date(transaction.block_time * 1000);
 
   return (
     <div className="flex flex-wrap items-center gap-3">
@@ -40,28 +40,28 @@ export const TransactionRow = ({ transaction }: TransactionRowProps) => {
       <div className="order-2 flex-1 md:flex-none">
         <Button variant="soft" size="1" color="gray" asChild>
           <a
-            href={`https://explorer.hiro.so/address/${transaction.sender}`}
+            href={`https://explorer.hiro.so/address/${transaction.sender_address}?chain=mainnet`}
             target="_blank"
             // Force the width to avoid content pushed on the right
             className="w-[108px]"
             rel="noreferrer"
           >
-            {transaction.sender.slice(0, 5)}...
-            {transaction.sender.slice(-5)}
+            {transaction.sender_address.slice(0, 5)}...
+            {transaction.sender_address.slice(-5)}
           </a>
         </Button>
       </div>
 
       <Tooltip
         content={`Block ${
-          transaction.blockHeight
-        } - ${timestamp.toUTCString()}`}
+          transaction.block_height
+        } - ${timestamp.toUTCString()}${transaction.tx_status === "success" ? "" : " (failed)"}`}
         suppressHydrationWarning
       >
         <Text
           className="order-3 md:order-4"
           size="2"
-          color="gray"
+          color={transaction.tx_status === "success" ? "gray" : "red"}
           suppressHydrationWarning
         >
           <TimeAgo date={timestamp} /> ago
@@ -76,7 +76,7 @@ export const TransactionRow = ({ transaction }: TransactionRowProps) => {
         asChild
       >
         <a
-          href={`https://explorer.hiro.so/txid/${transaction.txId}`}
+          href={`https://explorer.hiro.so/txid/${transaction.tx_id}?chain=mainnet`}
           target="_blank"
           rel="noreferrer"
           aria-label="View in Explorer"
@@ -85,20 +85,16 @@ export const TransactionRow = ({ transaction }: TransactionRowProps) => {
         </a>
       </IconButton>
 
-      <Text className="order-5 md:order-3 md:flex-1" as="div" size="2">
-        {transaction.action === "swap" ? (
-          <TransactionActionSwap transaction={transaction} />
-        ) : null}
-        {transaction.action === "add-liquidity" ? (
-          <TransactionActionAddLiquidity transaction={transaction} />
-        ) : null}
-        {transaction.action === "remove-liquidity" ? (
-          <TransactionActionRemoveLiquidity transaction={transaction} />
-        ) : null}
-        {transaction.action === "stackingdao-deposit" ||
-        transaction.action === "stackingdao-withdraw" ? (
+      <Text className="order-5 md:order-3 md:flex-1 truncate" as="div" size="2">
+        {transaction.protocol === "stackingdao" ? (
           <TransactionActionStackingDAO transaction={transaction} />
-        ) : null}
+        ) : transaction.protocol === "alex" ? (
+          <TransactionActionAlex transaction={transaction} />
+        ) : transaction.protocol === "velar" ? (
+          <TransactionActionVelar transaction={transaction} />
+        ) : (
+          <DefaultTransactionAction transaction={transaction} />
+        )}
       </Text>
     </div>
   );