+++ apps/web/src/lib/api.ts
@@ -1,3 +1,4 @@
+import type { ContractCallTransaction } from "@stacks/stacks-blockchain-api-types";
 import type { Protocol } from "./protocols";
 
 /**
@@ -23,6 +24,14 @@ export type ProtocolUsersRouteQuery = {
 };
 
 /**
+ * `/api/transactions`
+ */
+
+export type TransactionsRouteResponse = (ContractCallTransaction & {
+  protocol: Protocol;
+})[];
+
+/**
  * `/api/transactions/stats`
  */
 
@@ -34,3 +43,13 @@ export type TransactionStatsRouteResponse = {
 export type TransactionStatsRouteQuery = {
   protocol?: Protocol;
 };
+
+/**
+ * `/api/protocols/stackingdao`
+ */
+
+export type StackingDAOProtocolStatsResponse = {
+  month: string;
+  deposits: number;
+  withdrawals: number;
+}[];